+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/resource_properties/camera/flux/camera_settings_dialog_state.h
@@ -115,7 +115,7 @@ struct CombinedProperties
     CombinedValue defaultKeepCameraTimeSettingsState = CombinedValue::None;
     CombinedValue defaultRemoteArchiveSynchronizationEnabledState = CombinedValue::None;
 
-    int maxFps = 0;
+    float maxFps = 0.0;
 };
 NX_REFLECTION_INSTRUMENT(CombinedProperties,
     (isDtsBased)(isVirtualCamera)(isIoModule)(isArecontCamera)(supportsAudio)(supportsVideo)
