+++ vms/client/nx_vms_client_desktop/unit_tests/src/nx/vms/client/desktop/resource_properties/camera/flux/camera_settings_dialog_state_reducer_ut.cpp
@@ -35,7 +35,7 @@ QnScheduleTaskList makeSchedule(
     Qn::RecordingType recordingType,
     Qn::RecordingMetadataTypes metadataTypes,
     Qn::StreamQuality streamQuality = Qn::StreamQuality::normal,
-    int fps = 0,
+    float fps = 0.0,
     int bitrateKbps = 0
 )
 {
